+++ src/components/Header.tsx
@@ -5,7 +5,7 @@ import { useMediaQuery } from 'react-responsive'
 import { Mobile } from '@src/hooks/useScreenHook'
 import type { RootState } from '@src/store/config'
 import { useDispatch, useSelector } from 'react-redux'
-import { getUserInfo } from '@src/api/authApi'
+import { getUserInfo } from '@src/api/userApi'
 import { getCookieToken } from '@src/storage/Cookie'
 import { useEffect } from 'react'
 import { SET_INFO, DELETE_INFO } from '@src/store/slices/infoSlice'
