+++ src/components/ResetPassword/ResetPw.tsx
@@ -3,7 +3,7 @@ import { COLORS, FONT } from '@src/globalStyles'
 import React, { useState } from 'react'
 import { useNavigate } from 'react-router'
 import useInput from '@src/hooks/useInputHook'
-import { editUserPw } from '@src/api/authApi'
+import { editUserPw } from '@src/api/userApi'
 import { removeCookieToken } from '@src/storage/Cookie'
 import { useDispatch } from 'react-redux'
 import { DELETE_TOKEN } from '@src/store/slices/authSlice'
@@ -47,6 +47,8 @@ const ResetPw = () => {
       dispatch(DELETE_INFO())
       alert('비밀번호 변경에 성공했습니다. 다시 로그인 해주세요.')
       navigate(PATH.LOGIN)
+    } else {
+      alert('비밀번호 변경에 실패했습니다. 양식을 다시 확인해주세요.')
     }
   }
