+++ src/pages/EditPost.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useState } from 'react'
 import Write from './Write'
-import { getUserInfo } from '@src/api/authApi'
+import { getUserInfo } from '@src/api/userApi'
 import { useLocation } from 'react-router'
 import useModal from '@src/hooks/useModal'
