+++ src/pages/Join.tsx
@@ -2,7 +2,7 @@ import styled from 'styled-components'
 import { useNavigate } from 'react-router-dom'
 import { COLORS, FONT } from '@src/globalStyles'
 import React, { useEffect, useState } from 'react'
-import { checkDuplicateEmail, join } from '@src/api/authApi'
+import { checkDuplicateEmail, join } from '@src/api/userApi'
 import useInput from '@src/hooks/useInputHook'
 import PATH from '@src/constants/pathConst'
 import Modal from '@src/components/Modal'
