+++ src/pages/MyPage.tsx
@@ -7,7 +7,7 @@ import Inner from '@src/components/Inner'
 import PCBoardCard from '@src/components/MyPage/PCBoardCard'
 import { useNavigate } from 'react-router'
 import { Link } from 'react-router-dom'
-import { getMyPost, getWishlist, getMyReply } from '@src/api/authApi'
+import { getMyPost, getWishlist, getMyReply } from '@src/api/userApi'
 import { useDispatch, useSelector } from 'react-redux'
 import { SET_WISHLIST } from '@src/store/slices/wishlistSlice'
 import { RootState } from '@src/store/config'
