+++ src/pages/MyPageDetail.tsx
@@ -7,7 +7,7 @@ import { useLocation } from 'react-router'
 import { useMediaQuery } from 'react-responsive'
 import Inner from '@src/components/Inner'
 import Board from '@src/components/Board'
-import { getMyPost, getWishlist } from '@src/api/authApi'
+import { getMyPost, getWishlist } from '@src/api/userApi'
 import styled from 'styled-components'
 import MBoardList from '@src/components/MBoardList'
 import ReactPaginate from 'react-paginate'
