+++ src/pages/Profile.tsx
@@ -5,7 +5,7 @@ import styled from 'styled-components'
 import { COLORS, FONT } from '@src/globalStyles'
 import MBoardList from '@src/components/MBoardList'
 import { useState, useEffect } from 'react'
-import { getUserPost } from '@src/api/authApi'
+import { getUserPost } from '@src/api/userApi'
 import { useLocation } from 'react-router'
 import { useInView } from 'react-intersection-observer'
 import { useSelector } from 'react-redux'
